# mini-xrm-backend/src/gen/java/com/example/minixrm/backend/web/openapi/v1/api/ActivityApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.19.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.minixrm.backend.web.openapi.v1.api;

import com.example.minixrm.backend.web.openapi.v1.model.ActivityPageView;
import com.example.minixrm.backend.web.openapi.v1.model.ActivityView;
import com.example.minixrm.backend.web.openapi.v1.model.CreateOrUpdateActivityRequestView;
import com.example.minixrm.backend.web.openapi.v1.model.ErrorResponseView;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.19.0")
@Validated
@Controller
public interface ActivityApi {

    String PATH_CREATE_ACTIVITY = "/v1/activities";
    /**
     * POST /v1/activities : Create a new activity
     *
     * @param createOrUpdateActivityRequestView  (required)
     * @return No Content (status code 204)
     *         or Bad Request, e.g. validation failed for the request body or parameters (status code 400)
     *         or Conflict, e.g. trying to create an entity with a unique field value that already exists, or concurrent modification detected (status code 409)
     *         or Unprocessable Entity, e.g. an entity referenced by an ID in the request body is missing (status code 422)
     *         or Internal Server Error (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = ActivityApi.PATH_CREATE_ACTIVITY,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<Void> createActivity(
         @Valid @RequestBody CreateOrUpdateActivityRequestView createOrUpdateActivityRequestView
    );


    String PATH_DELETE_ACTIVITY = "/v1/activities/{activityId}";
    /**
     * DELETE /v1/activities/{activityId} : Delete an activity by ID
     *
     * @param activityId The ID of the activity (required)
     * @return No Content (status code 204)
     *         or Bad Request, e.g. validation failed for the request body or parameters (status code 400)
     *         or Not Found, e.g. trying to retrieve, update or delete an entity by an ID in the path does not exist (status code 404)
     *         or Conflict, e.g. trying to create an entity with a unique field value that already exists, or concurrent modification detected (status code 409)
     *         or Unprocessable Entity, e.g. an entity referenced by an ID in the request body is missing (status code 422)
     *         or Internal Server Error (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = ActivityApi.PATH_DELETE_ACTIVITY,
        produces = { "application/json" }
    )
    ResponseEntity<Void> deleteActivity(
        @NotNull @Min(value = 1L) @Max(value = 9223372036854775807L)  @PathVariable("activityId") Long activityId
    );


    String PATH_LIST_ACTIVITIES_BY_PARTNER_ID = "/v1/partners/{partnerId}/activities";
    /**
     * GET /v1/partners/{partnerId}/activities : Retrieve a list of activities by partner ID
     *
     * @param partnerId The ID of the partner (required)
     * @param page The page number to retrieve (0-based) (required)
     * @param pageSize The number of items per page (required)
     * @return Successful response (status code 200)
     *         or Bad Request, e.g. validation failed for the request body or parameters (status code 400)
     *         or Not Found, e.g. trying to retrieve, update or delete an entity by an ID in the path does not exist (status code 404)
     *         or Internal Server Error (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = ActivityApi.PATH_LIST_ACTIVITIES_BY_PARTNER_ID,
        produces = { "application/json" }
    )
    ResponseEntity<ActivityPageView> listActivitiesByPartnerId(
        @NotNull @Min(value = 1L) @Max(value = 9223372036854775807L)  @PathVariable("partnerId") Long partnerId,
        @NotNull @Min(value = 0)  @Valid @RequestParam(value = "page", required = true) Integer page,
        @NotNull @Min(value = 1)  @Valid @RequestParam(value = "pageSize", required = true) Integer pageSize
    );


    String PATH_LOAD_ACTIVITY = "/v1/activities/{activityId}";
    /**
     * GET /v1/activities/{activityId} : Retrieve an activity by ID
     *
     * @param activityId The ID of the activity (required)
     * @return Successful response (status code 200)
     *         or Bad Request, e.g. validation failed for the request body or parameters (status code 400)
     *         or Not Found, e.g. trying to retrieve, update or delete an entity by an ID in the path does not exist (status code 404)
     *         or Internal Server Error (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = ActivityApi.PATH_LOAD_ACTIVITY,
        produces = { "application/json" }
    )
    ResponseEntity<ActivityView> loadActivity(
        @NotNull @Min(value = 1L) @Max(value = 9223372036854775807L)  @PathVariable("activityId") Long activityId
    );


    String PATH_UPDATE_ACTIVITY = "/v1/activities/{activityId}";
    /**
     * PUT /v1/activities/{activityId} : Update an existing activity
     *
     * @param activityId The ID of the activity (required)
     * @param createOrUpdateActivityRequestView  (required)
     * @return No Content (status code 204)
     *         or Bad Request, e.g. validation failed for the request body or parameters (status code 400)
     *         or Not Found, e.g. trying to retrieve, update or delete an entity by an ID in the path does not exist (status code 404)
     *         or Conflict, e.g. trying to create an entity with a unique field value that already exists, or concurrent modification detected (status code 409)
     *         or Unprocessable Entity, e.g. an entity referenced by an ID in the request body is missing (status code 422)
     *         or Internal Server Error (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.PUT,
        value = ActivityApi.PATH_UPDATE_ACTIVITY,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<Void> updateActivity(
        @NotNull @Min(value = 1L) @Max(value = 9223372036854775807L)  @PathVariable("activityId") Long activityId,
         @Valid @RequestBody CreateOrUpdateActivityRequestView createOrUpdateActivityRequestView
    );

}