# mini-xrm-backend/src/gen/java/com/example/minixrm/backend/web/openapi/v1/api/PartnerApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.19.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.minixrm.backend.web.openapi.v1.api;

import com.example.minixrm.backend.web.openapi.v1.model.CreateOrUpdatePartnerRequestView;
import com.example.minixrm.backend.web.openapi.v1.model.ErrorResponseView;
import org.springframework.lang.Nullable;
import com.example.minixrm.backend.web.openapi.v1.model.PartnerPageView;
import com.example.minixrm.backend.web.openapi.v1.model.PartnerView;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.19.0")
@Validated
@Controller
public interface PartnerApi {

    String PATH_CREATE_PARTNER = "/v1/partners";
    /**
     * POST /v1/partners : Create a new partner
     *
     * @param createOrUpdatePartnerRequestView  (required)
     * @return No Content (status code 204)
     *         or Bad Request, e.g. validation failed for the request body or parameters (status code 400)
     *         or Conflict, e.g. trying to create an entity with a unique field value that already exists, or concurrent modification detected (status code 409)
     *         or Unprocessable Entity, e.g. an entity referenced by an ID in the request body is missing (status code 422)
     *         or Internal Server Error (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = PartnerApi.PATH_CREATE_PARTNER,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<Void> createPartner(
         @Valid @RequestBody CreateOrUpdatePartnerRequestView createOrUpdatePartnerRequestView
    );


    String PATH_DELETE_PARTNER = "/v1/partners/{partnerId}";
    /**
     * DELETE /v1/partners/{partnerId} : Delete a partner by ID
     *
     * @param partnerId The ID of the partner (required)
     * @return No Content (status code 204)
     *         or Bad Request, e.g. validation failed for the request body or parameters (status code 400)
     *         or Not Found, e.g. trying to retrieve, update or delete an entity by an ID in the path does not exist (status code 404)
     *         or Conflict, e.g. trying to create an entity with a unique field value that already exists, or concurrent modification detected (status code 409)
     *         or Unprocessable Entity, e.g. an entity referenced by an ID in the request body is missing (status code 422)
     *         or Internal Server Error (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = PartnerApi.PATH_DELETE_PARTNER,
        produces = { "application/json" }
    )
    ResponseEntity<Void> deletePartner(
        @NotNull @Min(value = 1L) @Max(value = 9223372036854775807L)  @PathVariable("partnerId") Long partnerId
    );


    String PATH_LOAD_PARTNER = "/v1/partners/{partnerId}";
    /**
     * GET /v1/partners/{partnerId} : Retrieve a partner by ID
     *
     * @param partnerId The ID of the partner (required)
     * @return Successful response (status code 200)
     *         or Bad Request, e.g. validation failed for the request body or parameters (status code 400)
     *         or Not Found, e.g. trying to retrieve, update or delete an entity by an ID in the path does not exist (status code 404)
     *         or Internal Server Error (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = PartnerApi.PATH_LOAD_PARTNER,
        produces = { "application/json" }
    )
    ResponseEntity<PartnerView> loadPartner(
        @NotNull @Min(value = 1L) @Max(value = 9223372036854775807L)  @PathVariable("partnerId") Long partnerId
    );


    String PATH_SEARCH_PARTNERS = "/v1/partners";
    /**
     * GET /v1/partners : Retrieve a list of partners
     *
     * @param page The page number to retrieve (0-based) (required)
     * @param pageSize The number of items per page (required)
     * @param partnerTagId The ID of the partner tag to filter by (optional)
     * @return Successful response (status code 200)
     *         or Bad Request, e.g. validation failed for the request body or parameters (status code 400)
     *         or Internal Server Error (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = PartnerApi.PATH_SEARCH_PARTNERS,
        produces = { "application/json" }
    )
    ResponseEntity<PartnerPageView> searchPartners(
        @NotNull @Min(value = 0)  @Valid @RequestParam(value = "page", required = true) Integer page,
        @NotNull @Min(value = 1)  @Valid @RequestParam(value = "pageSize", required = true) Integer pageSize,
        @Min(value = 1L) @Max(value = 9223372036854775807L)  @Valid @RequestParam(value = "partnerTagId", required = false) @Nullable Long partnerTagId
    );


    String PATH_UPDATE_PARTNER = "/v1/partners/{partnerId}";
    /**
     * PUT /v1/partners/{partnerId} : Update an existing partner
     *
     * @param partnerId The ID of the partner (required)
     * @param createOrUpdatePartnerRequestView  (required)
     * @return No Content (status code 204)
     *         or Bad Request, e.g. validation failed for the request body or parameters (status code 400)
     *         or Not Found, e.g. trying to retrieve, update or delete an entity by an ID in the path does not exist (status code 404)
     *         or Conflict, e.g. trying to create an entity with a unique field value that already exists, or concurrent modification detected (status code 409)
     *         or Unprocessable Entity, e.g. an entity referenced by an ID in the request body is missing (status code 422)
     *         or Internal Server Error (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.PUT,
        value = PartnerApi.PATH_UPDATE_PARTNER,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<Void> updatePartner(
        @NotNull @Min(value = 1L) @Max(value = 9223372036854775807L)  @PathVariable("partnerId") Long partnerId,
         @Valid @RequestBody CreateOrUpdatePartnerRequestView createOrUpdatePartnerRequestView
    );

}